restructure the bootstrap and results object
set up for any number of estimators now

Is not yet coordinated with frontend
