+++ backend/app/travel_times/estimateParameters.py
@@ -1,0 +1,69 @@
+from traveltimetools.utils import timeFormats
+from random import choices
+from numpy import percentile, quantile
+import polars
+
+# Q: Is this the ideal way to do this? 
+# A: It is the way we currently do it.
+def mean_daily_mean(observations):
+    """Takes a list of DynamicBins and returns
+    the *average* of the *daily averages*"""
+
+    return polars.DataFrame({
+        'date': [bin.singleDate for bin in observations],
+        'travelTime': [bin.travelTime for bin in observations]
+    }).group_by('date').agg(
+        polars.col('travelTime').mean().alias('travelTime')
+    ).select(
+        polars.col('travelTime').mean()
+    ).item()
+
+def median(observations):
+    return quantile(
+        [obs.travelTime for obs in observations],
+        0.5
+    )
+
+# this is a quite low, but quick
+resamples = 100
+
+# 95% confidence interval
+pctLower = 2.5
+pctUpper = 97.5
+
+functions = {
+    'mean': mean_daily_mean,
+    'median': median
+}
+
+def estimateParameters(sample):
+    """returns bootstrapped confidence intervals for the sample"""
+    if len(sample) == 0:
+        return None
+    data = {}
+    
+    for funcName, func in functions.items():
+        data[funcName] = {
+            'estimate': timeFormats(func(sample)),
+            'bootstrapDistribution': []
+        }
+
+    for i in range(0, resamples):
+        # resample with replacement
+        bootstrapSample = choices( sample, k = len(sample) )
+        # estimate functions from resampled distribution
+        for funcName, func in functions.items():
+            data[funcName]['bootstrapDistribution'].append(func(bootstrapSample))
+
+    for estimate in data.values():
+        lowerCI, upperCI = percentile(
+            estimate['bootstrapDistribution'],
+            [pctLower, pctUpper]
+        )
+        estimate['ci'] = {
+            'lower': timeFormats(lowerCI),
+            'upper': timeFormats(upperCI)
+        }
+        del estimate['bootstrapDistribution']
+
+    return data
